+++ index.html
@@ -1,3 +1,4 @@
+<!DOCTYPE html>
 <html>
   <head>
     <link rel="icon" href="data:;base64,iVBORw0KGgo=">
@@ -5,6 +6,6 @@
   </head>
   <body>
     <main id="minideck"></main>
-    <script src="./index.js"></script>
+    <script type="module" src="./index.jsx"></script>
   </body>
 </html>
